[Feature #162946769] add view own records and count
