+++ app/api/v2/routes_v2.py
@@ -1,7 +1,7 @@
 
 from ..v2.views.users_view import UsersView, OneUser, MakeAdmin
 from ..v2.views.incidents_view import IncidentViews, OneIntervention,\
-    EditComment, EditLocation, ChangeStatus
+    EditComment, EditLocation, ChangeStatus, UserIncidents, IncidentsCount
 from flask_restful import Api, Resource
 from flask import Blueprint
 
@@ -18,3 +18,5 @@ api.add_resource(OneIntervention, '<type>/<incident_id>')
 api.add_resource(EditComment, '<type>/<incident_id>/comment')
 api.add_resource(EditLocation, '<type>/<incident_id>/location')
 api.add_resource(ChangeStatus, '<type>/<incident_id>/status')
+api.add_resource(UserIncidents, 'user/<type>')
+api.add_resource(IncidentsCount, 'user/<type>/count')
